backend-merged/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.2
email-validator==2.1.0.post1

# AI Integration
google-genai==1.68.0

# Biometric Processing
opencv-python==4.9.0.80
face-recognition==1.3.0
librosa==0.10.1
numpy==1.26.4
pillow==10.2.0

# Utilities
python-multipart==0.0.20
httpx==0.28.1
aiofiles==23.2.1
python-dotenv==1.0.1

# Performance & Monitoring
slowapi==0.1.9
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0
